shoe-tracker/requirements.txt
firebase-admin>=5.0.0
google-cloud-firestore>=2.0.0
requests>=2.28.0
python-dateutil>=2.8.2
